Moved code to internal packages
Those utility classes don't need to be public.
