+++ src/main/groovy/org/mockito/release/internal/util/MultiMap.java
@@ -1,4 +1,4 @@
-package org.mockito.release.util;
+package org.mockito.release.internal.util;
 
 import java.util.*;
